backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
spotipy==2.23.0
openai==2.6.0
pillow==10.2.0
python-dotenv==1.0.0
werkzeug==3.0.1